requirements.txt
aiohttp>=3.9
beautifulsoup4>=4.12
pydantic>=2.8
python-dotenv>=1.0
openai>=1.40
tenacity>=8.2
pytest>=8.2
pytest-asyncio>=0.23
black>=24.4
ruff>=0.5


